Added cpu timer
Well, it doesn't work *at all*. When I include these files with the
Makefile everything crashes. I *think* this is due to the fact that the
bootloader only handles direct ext2 pointer and the kernel is too big to
fit into them. Therefore the kernel taps into the void as it tries to
read some data an crashes.

It could be something completely different though - let's see!
(this will take some time ig)